pyproject.toml
[tool.poetry]
name = "flashrecord"
version = "0.3.3"
description = "Fast, simple screen capture and animated GIF recording with intelligent compression"
authors = ["Flamehaven"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Flamehaven/flashrecord"
repository = "https://github.com/Flamehaven/flashrecord"
keywords = ["screenshot", "screen-capture", "gif", "screen-recording", "compression"]
packages = [{include = "flashrecord", from = "src"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.8"
pillow = ">=9.0.0"
imageio = ">=2.0.0"
numpy = ">=1.20.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.0.0"
pytest-xdist = ">=3.0.0"
pytest-rerunfailures = ">=12.0.0"
pytest-timeout = ">=2.1.0"
ruff = ">=0.1.0"
black = ">=23.0.0"
mypy = ">=1.0.0"
sphinx = ">=7.0.0"
sphinx-rtd-theme = ">=2.0.0"
myst-parser = ">=2.0.0"

[tool.poetry.scripts]
flashrecord = "flashrecord.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=flashrecord",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--reruns=2",
    "--reruns-delay=1",
    "--timeout=300",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "flaky: Tests that may fail intermittently",
]
timeout = 300
timeout_method = "thread"

[tool.coverage.run]
source = ["src/flashrecord"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
